+++ franka_robot_server/src/gripper_control.cpp
@@ -19,6 +19,7 @@
 FrankaRobotRPCServiceImpl::~FrankaRobotRPCServiceImpl() {
     stopGripperWorkerThread();
     stopVacuumGripperWorkerThread();
+    stopMotionWorkerThread();
 }
 
 // ============================================================================